requirements.txt
Flask==2.0.1
ChatterBot2
chatterbot-corpus
scikit-learn
spacy 
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0.tar.gz


# This is the correct requirement file